LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicyviolation-basics-expected.txt

Check that a SecurityPolicyViolationEvent handler exists, and that events can be created.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof document.onsecuritypolicyviolation is "object"
PASS typeof SecurityPolicyViolationEvent is "function"
PASS typeof window.e is "object"
PASS window.e.documentURI is "documentURIValue"
PASS window.e.referrer is "referrerValue"
PASS window.e.blockedURI is "blockedURIValue"
PASS window.e.violatedDirective is "violatedDirectiveValue"
PASS window.e.effectiveDirective is "effectiveDirectiveValue"
PASS window.e.originalPolicy is "originalPolicyValue"
PASS window.e.sourceURL is "sourceURLValue"
PASS window.e.lineNumber is 1
PASS successfullyParsed is true

TEST COMPLETE
